refactor: Modularizes ECS inference logic
Splits a large monolithic inference file into focused modules for cluster inference, analysis loading, performance tracking, task scoring, and main orchestration.
Introduces a central module to re-export interfaces and main functions for cleaner imports.
Keeps a thin compatibility layer for existing imports.
Improves maintainability, clarity, and enables easier testing and future extension.
